refactor: schemas 폴더 삭제, dtos와 병합
