+++ src/dtos/review.dto.ts
@@ -1,12 +1,25 @@
 import z from "zod";
-import {
-  CreateReviewSchema,
-  ReviewIdParamsSchema,
-  UpdateReviewschema,
-} from "../schemas/review.schema";
+
+// 리뷰 생성 DTO 및 zod 스키마
+export const CreateReviewSchema = z.object({
+  estimateId: z.string().uuid(),
+  rating: z.number().int().min(1).max(5),
+  content: z.string().min(10),
+});
 
 export type CreateReviewDto = z.infer<typeof CreateReviewSchema>;
 
+// 리뷰 수정 DTO 및 zod 스키마
+export const UpdateReviewschema = z.object({
+  rating: z.number().int().min(1).max(5).optional(),
+  content: z.string().min(1).optional(),
+});
+
 export type UpdateReviewDto = z.infer<typeof UpdateReviewschema>;
+
+// 리뷰 ID 파라미터 DTO 및 zod 스키마
+export const ReviewIdParamsSchema = z.object({
+  reviewId: z.string().uuid(),
+});
 
 export type ReviewIdParamsDto = z.infer<typeof ReviewIdParamsSchema>;
